+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@ Release Notes
 =============
 ## vNext
 * SQL Azure: Clean up Entra ID authentication support.
+* Web App: Support automatic creation of Workspace Based App Insights & Log Analytics.
+* Functions: Support automatic creation of Workspace Based App Insights & Log Analytics.
 * Az: Update `ad` commands to work with latest (breaking) structure.
 * PostgreSQL: Fix a number of issues around the introduction of Flexible Servers.